Refactor: Move SandboxTarget to fixtures, rename to ListTarget
